New coreutils, ACPI improvements,SMP,removed  namespace, new syscalls

+++ kernel/arch/amd64/pci.h
@@ -4,7 +4,7 @@
 #include <arch/amd64/types.h>
 
 //MI PCI interface
-uint32 arch_pci_read(uint8 bus, uint8 dev, uint8 func, uint8 offset, uint8 size);
-void arch_pci_write(uint8 bus, uint8 dev, uint8 func, uint8 offset, uint8 size, uint32 value);
+uint32 arch_pci_read(uint8 bus, uint8 dev, uint8 func, uint16 offset, uint8 size);
+void arch_pci_write(uint8 bus, uint8 dev, uint8 func, uint16 offset, uint8 size, uint32 value);
 
 #endif
